Move ListView to common from components
