+++ assets/js/components/ClusterDetails/HanaClusterDetails.jsx
@@ -6,7 +6,7 @@ import { RUNNING_STATES } from '@state/lastExecutions';
 import PageHeader from '@components/PageHeader';
 import BackButton from '@components/BackButton';
 
-import ListView from '@components/ListView';
+import ListView from '@common/ListView';
 import Table from '@components/Table';
 import Tooltip from '@components/Tooltip';
 import ClusterNodeLink from '@components/ClusterDetails/ClusterNodeLink';
