+++ assets/js/components/HostDetails/SaptuneSummary.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { isVersionSupported } from '@lib/saptune';
 
 import Button from '@common/Button';
-import ListView from '@components/ListView';
+import ListView from '@common/ListView';
 import { SaptuneVersion, SaptuneTuningState } from '@components/SaptuneDetails';
 
 function SaptuneSummary({
