+++ assets/js/components/SapSystemDetails/GenericSystemDetails.jsx
@@ -7,7 +7,7 @@ import {
 
 import { APPLICATION_TYPE, getEnsaVersionLabel } from '@lib/model/sapSystems';
 
-import ListView from '@components/ListView';
+import ListView from '@common/ListView';
 import Table from '@components/Table';
 import PageHeader from '@components/PageHeader';
 import DeregistrationModal from '@components/DeregistrationModal';
